Save player ship on jump
Here we encounter the old save data problems again. We
cannot use Unity serialization for saving; that's an editor
thing only. I have opted for a super minimal binary serialization
system that borrows from the network serializer.

This worked first time without issues, which is encouraging!
Suggests we're homing in on a sound infrastructure.